Cleanup transient id
