+++ hapi-fhir-jpaserver-model/src/main/java/ca/uhn/fhir/jpa/model/entity/ResourceHistoryTable.java
@@ -22,7 +22,26 @@ package ca.uhn.fhir.jpa.model.entity;
 import ca.uhn.fhir.jpa.model.dao.JpaPid;
 import ca.uhn.fhir.model.primitive.IdDt;
 import ca.uhn.fhir.rest.api.Constants;
-import jakarta.persistence.*;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.ForeignKey;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Index;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.Lob;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
+import jakarta.persistence.SequenceGenerator;
+import jakarta.persistence.Table;
+import jakarta.persistence.Transient;
+import jakarta.persistence.UniqueConstraint;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
 import org.hibernate.Length;
@@ -111,6 +130,13 @@ public class ResourceHistoryTable extends BaseHasResource implements Serializabl
 
 	@Transient
 	private transient ResourceHistoryProvenanceEntity myNewHistoryProvenanceEntity;
+	/**
+	 * This is stored as an optimization to avoid needing to query for this
+	 * after an update
+	 */
+	@Transient
+	// TODO MB forced_id delete this in step 3
+	private transient String myTransientForcedId;
 
 	/**
 	 * Constructor
@@ -300,5 +326,13 @@ public class ResourceHistoryTable extends BaseHasResource implements Serializabl
 			myNewHistoryProvenanceEntity = new ResourceHistoryProvenanceEntity();
 		}
 		return myNewHistoryProvenanceEntity;
+	}
+
+	public String getTransientForcedId() {
+		return myTransientForcedId;
+	}
+
+	public void setTransientForcedId(String theTransientForcedId) {
+		myTransientForcedId = theTransientForcedId;
 	}
 }
